Moved routes for products and articles into their route files and used express' router to organize the routing.
