+++ views/article.hbs
@@ -6,3 +6,4 @@
 <a class="edit" href="/articles/{{title}}/edit">Edit Article</a>
 <br>
 <a class="redirect" href="/articles"><i class="fas fa-arrow-circle-left"></i> Back to Articles</a>
+<br> {{> footer}}
