+++ views/index.hbs
@@ -18,5 +18,4 @@
 <a class="add" href="/articles/new"><i class="fas fa-plus"></i> Add a new article</a>
 {{/if}}
 
-<br><br>
-<a class="redirect" href="/"><i class="fas fa-arrow-circle-left"></i> HOME</a>
+<br><br> {{!-- <a class="redirect" href="/"><i class="fas fa-arrow-circle-left"></i> HOME</a> --}} {{> footer}}
