+++ views/new.hbs
@@ -53,7 +53,7 @@
 </form>
 <br>
 <a class="redirect" href="/articles"><i class="fas fa-arrow-circle-left"></i> Back to Articles</a>
-{{/if}} {{#if addingArticleError}}
+<br> {{> footer}} {{/if}} {{#if addingArticleError}}
 <h1>Problem with submitting new article. Please fill in all fields and try again.</h1>
 <form action="/articles" method="post">
   <label>Title:<br>
@@ -72,4 +72,4 @@
 </form>
 <br>
 <a class="redirect" href="/articles"><i class="fas fa-arrow-circle-left"></i> Back to Articles</a>
-{{/if}}
+<br> {{> footer}} {{/if}}
