+++ views/partials/footer.hbs
@@ -1,0 +1,1 @@
+<a class="redirect" href="/"><i class="fas fa-arrow-circle-left"></i> HOME</a>
