+++ views/product.hbs
@@ -7,3 +7,4 @@
 
 <br>
 <a class="redirect" href="/products"><i class="fas fa-arrow-circle-left"></i> Back to Products</a>
+<br> {{> footer}}
